requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.20.0
pyjwt==2.10.1
python-decouple==3.8
passlib==1.7.4
celery==5.6.2
redis==7.1.1
httpx==0.28.1
python-multipart==0.0.20
psycopg[pool]==3.3.2
nanoid==2.0.0
sqlalchemy==2.0.47
python-slugify==8.0.1
requests==2.32.4
pydantic==2.12
faker==18.3.0
pytest==7.4.0
pytest-asyncio==0.20.3